Refactor EPOS4Bus to a separate module
Moved the EPOS4Bus class to a new file (epos4bus.py) to improve maintainability by modularizing the code. Updated imports accordingly in related files and removed redundant code in epos4.py.

This is also helping hilight the questionable class model of having both EthercatBus and EPOS4Bus. I'm thinking Ethercat bus should be subsumed into EPOS4Bus
