+++ cooethercat/__init__.py
@@ -1,4 +1,5 @@
-from .epos4 import EPOS4Bus, EPOS4Motor
+from .epos4 import EPOS4Motor
+from .epos4bus import EPOS4Bus
 from .helpers import StatuswordStates
 
 __all__ = [EPOS4Bus, StatuswordStates, EPOS4Motor]
